\name{ppgmmga-package}
\alias{ppgmmga-package}
\docType{package}

\title{Projection pursuit based on Gaussian Mixtures Models and Evolutionary Algorithms for data visualisation}

\description{
An R package implementing a Projection Pursuit algorithm based on finite Gaussian mixtures for density estimation and Genetic Algorithms to maximise the approximate Negentropy index for Gaussian mixtures. The \emph{ppgmmga} algorithm provides a method to visualise high-dimensional data in a lower-dimensional space.
}

\details{
An introdunction to \code{ppgmmga} package is presented in the \href{../doc/index.html}{vignette}. 
}


\author{
Serafini A. \email{srf.alessio@gmail.com}

Scrucca L. \email{luca.scrucca@unipg.it}

}


\seealso{\code{\link{ppgmmga}}, \code{\link{plot.ppgmmga}}, \code{\link{ppgmmga-class}}, \code{\link{ppgmmga.options}}, \code{\link{summary.ppgmmga}}, \code{\link{EntropyGMM}}, \code{\link{NegentropyGMM}}}