\name{ppgmmga-package}
\alias{ppgmmga-package}
\docType{package}

\title{Projection pursuit based on Gaussian mixtures and evolutionary algorithms for data visualisation}

\description{
An R packege implementing a projection pursuit algorithm based on finite Gaussian mixtures for density estimation, and genetic algorithms to maximise the approximate Negentropy index for Gaussian mixtures. The ppgmmga algorithm implemented in the package provides a method to visualise high-dimensional data. In adddition the package provide an anomaly detection algorithm based on the ppgmmga algorithm and Gaussian mixture with Uniform component to detect the anomalies in a dataset. 
}

\details{
For a quick intro to ppgmmga package see the \href{}{vignette}. 
}

\references{
%% ~put references to the literature/web site here ~
}
\author{
Serafini A. \email{srf.alessio@gmail.com}

Scrucca L. \email{luca.scrucca@unipg.it}
}


\seealso{\code{\link{ppgmmga}},\code{\link{plot.ppgmmga}},\code{\link{ppgmmga-class}},\code{\link{ppgmmga.options}},\code{\link{ppgmmga.options.restore}},\code{\link{summary.ppgmmga}},\code{\link{EntropyGMM}},\code{\link{NegentropyGMM}}}