\name{ppgmmga-package}
\alias{ppgmmga-package}
\docType{package}

\title{Projection pursuit based on Gaussian mixtures and evolutionary algorithms for data visualisation}

\description{
An R package implementing a Projection Pursuit (PP) algorithm based on finite Gaussian Mixture Models (GMMs) for density estimation using Genetic Algorithms (GAs) to maximise an approximated negentropy index. The \pkg{ppgmmga} algorithm provides a method to visualise high-dimensional data in a lower-dimensional space.
}

\details{
An introduction to \pkg{ppgmmga} package is provided in the accompaying vignette \href{../doc/ppgmmga.html}{A quick tour of ppgmmga}. 
}

\author{
Serafini A. \email{srf.alessio@gmail.com}\cr
Scrucca L. \email{luca.scrucca@unipg.it}
}

\seealso{\code{\link{ppgmmga}}, \code{\link{plot.ppgmmga}}, \code{\link{ppgmmga-class}}, \code{\link{ppgmmga.options}}, \code{\link{summary.ppgmmga}}}

\references{
Scrucca, L. and Serafini, A. (2019) Projection pursuit based on Gaussian mixtures and evolutionary algorithms. *Journal of Computational and Graphical Statistics*, DOI: 10.1080/10618600.2019.1598871 
}